requirements.txt
asyncio==3.4.3
python-telegram-bot==20.4
requests==2.31.0